// src/app/services/api/mtc/model/televisitaListaRichiesteAssistitoResponse.ts
/**
 * Api Documentation
 * Api Documentation
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { TelevisitaResponse } from './televisitaResponse';


export interface TelevisitaListaRichiesteAssistitoResponse { 
    /**
     * Televisite
     */
    televisite?: Array<TelevisitaResponse>;
}