// src/app/services/api/mtc/model/televisitaResponse.ts
/**
 * Api Documentation
 * Api Documentation
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { TelevisitaAssistitoResponse } from './televisitaAssistitoResponse';
import { TelevisitaDurataResponse } from './televisitaDurataResponse';
import { TelevisitaPartecipanteResponse } from './televisitaPartecipanteResponse';


export interface TelevisitaResponse { 
    /**
     * Altri partecipanti della televisita
     */
    altriPartecipanti?: Array<TelevisitaPartecipanteResponse>;
    /**
     * Assistito
     */
    assistito?: TelevisitaAssistitoResponse;
    /**
     * Data (yyyy-MM-dd)
     */
    data?: string;
    /**
     * Durata
     */
    durata?: TelevisitaDurataResponse;
    /**
     * ID televisita
     */
    id?: string;

    motivazioneMedico?: string;
    /**
     * Campo motivazione
     */
    motivazione?: string;
    /**
     * Campo note
     */
    note?: string;
    /**
     * Campo note (medico)
     */
    noteMedico: string;
    /**
     * Ora (hh:mm a)
     */
    ora?: string;
    /**
     * Stato
     */
    stato?: TelevisitaResponse.StatoEnum;
}
export namespace TelevisitaResponse {
    export type StatoEnum = 'PRESENTATA' | 'RITIRATA' | 'RESPINTA' | 'PRESA_IN_CARICO' | 'RIFIUTATA' | 'ACCETTATA' | 'SCADUTA' | 'IN_CORSO' | 'EFFETTUATA' | 'INTERROTTA';
    export const StatoEnum = {
        PRESENTATA: 'PRESENTATA' as StatoEnum,
        RITIRATA: 'RITIRATA' as StatoEnum,
        RESPINTA: 'RESPINTA' as StatoEnum,
        PRESAINCARICO: 'PRESA_IN_CARICO' as StatoEnum,
        RIFIUTATA: 'RIFIUTATA' as StatoEnum,
        ACCETTATA: 'ACCETTATA' as StatoEnum,
        SCADUTA: 'SCADUTA' as StatoEnum,
        INCORSO: 'IN_CORSO' as StatoEnum,
        EFFETTUATA: 'EFFETTUATA' as StatoEnum,
        INTERROTTA: 'INTERROTTA' as StatoEnum
    };
}